+++ packages/react/src/index.ts
@@ -14,7 +14,7 @@ export { default as Radio } from './components/Radio';
 export { BasicTab, BorderTab } from './components/Tab';
 export { ColumnTable, RowTable, BasicTable } from './components/Table';
 export * from './components/Text';
-export { default as Icon } from './components/Icon';
+export * from './components/Icon';
 export { default as Slider } from './components/Slider';
 export { default as cosStyle, CosStyle } from './styles/GlobalStyle';
 export type { Theme, Palette } from './styles/theme';
